modularization: extract HTML generator
